bayesian-calculator/requirements.txt
pandas==2.1.4
scipy==1.11.4
google-cloud-bigquery==3.14.0
numpy==1.26.2
pytz==2023.3
